Nuevos cambios

+++ PeluqueriAPP/PeluqueriAPP/Servicio.cs
@@ -4,16 +4,13 @@
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public string descripcion { get; set; } // añadida descripción
+        public string descripcion { get; set; }
         public int duracion { get; set; }
-        public double precio { get; set; }
-
-        // Solo mantenemos la referencia al tipo de servicio para mostrar su nombre
+        public decimal precio { get; set; } // <-- de double a decimal
         public TipoServicio tipoServicio { get; set; }
-
-        // Propiedad calculada para facilitar la visualización en el DataGridView
         public string tipoServicioNombre => tipoServicio?.nombre ?? "";
     }
+
 
     public class TipoServicio
     {
